+++ client/orm/filter/prometheus/filter.go
@@ -39,11 +39,12 @@ type FilterChainBuilder struct {
 	RunMode    string
 }
 
-var summaryVec prometheus.ObserverVec
-var initSummaryVec sync.Once
+var (
+	summaryVec     prometheus.ObserverVec
+	initSummaryVec sync.Once
+)
 
 func (builder *FilterChainBuilder) FilterChain(next orm.Filter) orm.Filter {
-
 	initSummaryVec.Do(func() {
 		summaryVec = prometheus.NewSummaryVec(prometheus.SummaryOpts{
 			Name:      "beego",